% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCoefSD1.R
\name{calcCoefSD1}
\alias{calcCoefSD1}
\title{Calculate coefficients for SD1 method.}
\usage{
calcCoefSD1(data)
}
\arguments{
\item{data}{a list resulting from crMatrixSD function.}
}
\value{
a three column matrix, with parameters alpha, beta and gamma.
}
\description{
Calculate coefficients for SD1 method.
}
\examples{

}
