% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLambdas.R
\name{calcLambdas}
\alias{calcLambdas}
\title{Calculate a Holgate's Poisson parameters for a soccer match.}
\usage{
calcLambdas(coef, home, visitor, local = 1)
}
\arguments{
\item{coef}{a three column matrix resulting from calcCoefSD1 function.}

\item{home}{home team name.}

\item{visitor}{visitor team name.}

\item{local}{indicator for local factor to the match. Must be 0 or 1
(default).}
}
\value{
a vector of lenght three containing the Holgate's Poisson parameters
for the match.
}
\description{
Calculate a Holgate's Poisson parameters for a soccer match.
}
