% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMatProb.R
\name{calcMatProb}
\alias{calcMatProb}
\title{Calculate probabilities for a particular soccer match.}
\usage{
calcMatProb(lambdas, dim.matrix = 11)
}
\arguments{
\item{lambdas}{a vector of lenght three containing the Holgate's Poisson
parameters for the match (resulting from calcLambdas function).}

\item{dim.matrix}{matrix (square) dimension to use to calculate all probabilities.}
}
\value{
a list with two elements, Probs and Matrix.
}
\description{
Calculate probabilities for a particular soccer match.
}
