% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crMatrixSD.R
\name{crMatrixSD}
\alias{crMatrixSD}
\title{Creat a Data Matrix from a XLS File.}
\usage{
crMatrixSD(xlsFile, ...)
}
\arguments{
\item{xlsFile}{the spreadsheet file name to be loaded.}

\item{...}{additional parameters for read.xls function (gdata package).}
}
\value{
A list with two vectors: sum and dif; and two matrices: matS and matT.
}
\description{
This function creates a data matrix from a spreadsheet file containing
soccer matches results.
}
\examples{

}
